Fix --expand mode for empty attributes
If an attribute appears in sample_attribute but has no value
associated with it, it is as good as non-existent, we
discard this attribute. This only effects it on a record level
and not the entire run table. If any other sample in the same
run table has that attribute, the DataFrame will still have
all the relevant columns with the record with missing value
for an attribute as an NaN.

Closes #11