Introduce a fallback option to websearch
SRAdb seems to be lagging in some cases (https://github.com/seandavi/SRAdb/issues/24)
or missing datasets in other cases (https://github.com/seandavi/SRAdb/issues/23).
This commit introduces a fallback mechanism to deal with samples
that are not in the database or if the database is not required at all.
The detailed and expand tags are not supported yet.